requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.0.1
model-bakery==1.5.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2021.3
redis==4.3.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
